Changed name of folder to "simulator"
