+++ src/simulator/Compare.java
@@ -1,4 +1,4 @@
-package basic;
+package simulator;
 
 // Basic Imports
 import javafx.stage.Stage;
@@ -19,16 +19,16 @@ import javafx.geometry.Pos;
 import javafx.scene.layout.GridPane;
 import javafx.geometry.Insets;
 
-public class MyList {
-    public static void MyListScreen (Stage primaryStage, ArrayList<Games> GamesList) {
-        GridPane myListGrid = new GridPane();
-        myListGrid.setVgap(10);
-        myListGrid.setHgap(10);
-        myListGrid.setGridLinesVisible(false);
-        myListGrid.setPadding(new Insets(25, 25, 25, 25));
+public class Compare {
+    public static void CompareScreen (Stage primaryStage, ArrayList<Games> GamesList) {
+        GridPane compareGrid = new GridPane();
+        compareGrid.setVgap(10);
+        compareGrid.setHgap(10);
+        compareGrid.setGridLinesVisible(false);
+        compareGrid.setPadding(new Insets(25, 25, 25, 25));
 
         Button homeMenu = new Button();
-        myListGrid.add(homeMenu, 0, 0);
+        compareGrid.add(homeMenu, 0, 0);
         homeMenu.setText("Back");
         homeMenu.setMaxSize(100, 50);
         homeMenu.setOnAction(new EventHandler<ActionEvent>() {
@@ -40,7 +40,7 @@ public class MyList {
         });
 
         primaryStage.setWidth(601);
-        primaryStage.setScene(new Scene(myListGrid));
+        primaryStage.setScene(new Scene(compareGrid));
         primaryStage.show();
     }
 }
