+++ src/simulator/Games.java
@@ -1,4 +1,4 @@
-package basic;
+package simulator;
 
 import java.io.*;
 import java.util.*;
